crawling/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
cx_Oracle==8.3.0
oracledb==2.0.1
python-multipart==0.0.6
aiofiles==23.2.1


